generateattcks/requirements.txt
requests
pyyaml==5.1.0
pygithub
xlrd
markdown
googletrans